New PC Comand structure

+++ Device/modules/Dimensions.h
@@ -19,7 +19,7 @@ struct Calibrate
 
 struct Timing
 {
-    uint16_t tx; //Time between sending readings in ms
+    uint32_t tx; //Time between sending readings in ms
     uint16_t samp; //How often to sample the sensor in ms
     uint16_t reads; //how many reads of the sensor to average per sample
 };
